# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Friendtrip",
        "description": "Friendtrip api document (internal use only)",
        "contact": {
            "email": "chaniwat.meranote@gmail.com"
        },
        "version": "0.0.1"
    },
    "host": "friendtrip.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/authentication": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get authenticate user information",
                "description": "Get user information of given token",
                "operationId": "getAuthUserInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    }
                }
            },
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authentication (Request token)",
                "description": "Request token for user auth needed APIs",
                "operationId": "authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User identity",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AuthenticationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request token",
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    "401": {
                        "description": "Invalid credential",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "Get all event information",
                "description": "Get all event information",
                "operationId": "getAllEvents",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "show_per_page",
                        "in": "query",
                        "description": "Set show events per page",
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page",
                        "type": "integer"
                    },
                    {
                        "name": "full_event",
                        "in": "query",
                        "description": "Show full event or not?",
                        "type": "boolean"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search event by name",
                        "type": "string"
                    },
                    {
                        "name": "place",
                        "in": "query",
                        "description": "Search event by place",
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter event by type",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Events information",
                        "schema": {
                            "$ref": "#/definitions/AllEvents"
                        }
                    },
                    "400": {
                        "description": "Invalid token"
                    },
                    "401": {
                        "description": "Token expired"
                    }
                }
            },
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Create new event",
                "description": "Create new event into database (owner is current authentication user (token))",
                "operationId": "newEvent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Event information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/EventBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New event has been created"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "422": {
                        "description": "Invalid parameters"
                    }
                }
            }
        },
        "/events/{event_id}": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "Get event information",
                "description": "Get information of event {event_id}",
                "operationId": "getEventInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event information",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "400": {
                        "description": "Invalid token"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "event"
                ],
                "summary": "Update event information",
                "description": "Update event information of {event_id} (Need authentication token for self update or admin token for update to any event)",
                "operationId": "updateEvent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Event information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/EventBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update event information successful"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired or no permission"
                    },
                    "422": {
                        "description": "Invalid parameters"
                    }
                }
            }
        },
        "/events/{event_id}/join": {
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Join event",
                "description": "Join to event",
                "operationId": "joinEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Join event success"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "404": {
                        "description": "No event found"
                    }
                }
            }
        },
        "/events/{event_id}/leave": {
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Leave event",
                "description": "Leave to event",
                "operationId": "leaveEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "leave event success"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "404": {
                        "description": "No event found or not joined to event"
                    }
                }
            }
        },
        "/events/{event_id}/cancel": {
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Cancel event",
                "description": "Cancel the event",
                "operationId": "cancelEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cancel event success"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired or no permission (not an owner of event)"
                    },
                    "404": {
                        "description": "No event found"
                    }
                }
            }
        },
        "/images/event": {
            "post": {
                "tags": [
                    "image"
                ],
                "summary": "Upload new event image",
                "description": "Upload new image to server",
                "operationId": "newEvent",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Event information",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload successful"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "422": {
                        "description": "Invalid file type"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AllEvents": {
            "required": [
                "events",
                "pagination"
            ],
            "properties": {
                "events": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Event"
                    }
                },
                "pagination": {
                    "description": "",
                    "$ref": "#/definitions/Pagination"
                }
            },
            "type": "object"
        },
...
        "AuthenticationInfo": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "",
                    "type": "string"
                },
                "password": {
                    "description": "",
                    "type": "string"
                },
                "get_info": {
                    "description": "",
                    "type": "boolean",
                    "default": false
                }
            },
            "type": "object"
        },
        "Event": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer"
                },
                "owner": {
                    "description": "",
                    "$ref": "#/definitions/User"
                },
                "participant_count": {
                    "description": "",
                    "type": "integer"
                },
                "join_status": {
                    "description": "",
                    "type": "string",
                    "default": "owner or boolean (true = join, false = not join yet) | if sent token for checking status"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/EventBody"
                }
            ]
        },
...
        "Token": {
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "",
                    "type": "string"
                },
                "user": {
                    "description": "",
                    "$ref": "#/definitions/User"
                }
            },
            "type": "object"
        },
...
        "User": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/UserBody"
                }
            ]
        },
...
    },
...
}